pyproject.toml
[tool.poetry]
name = "jwt-fastapi"
version = "0.1.0"
description = "Fastapi jwt auth extension"
authors = ["Abel Alazar Dirar <aadlinux358@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "jwt_fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.100.1"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
ruff = "^0.0.282"
black = "^23.7.0"
jedi = "^0.19.0"
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
